requirements.txt
requests
SQLAlchemy
python-dotenv
PyYAML
pydantic